feat: update decimal precision in material calculations and UI components
- Added migration to change Float columns to Numeric(10, 2) in relevant tables for precise decimal handling.
- Updated calculations in project creation and material requirement updates to use Decimal for accuracy.
- Introduced FormattedNumberInput component for consistent number formatting across forms.
- Replaced standard select inputs with SearchableSelect for improved user experience in selecting materials and suppliers.
- Enhanced ProjectStageView and Purchases components to handle decimal precision and display correctly.

+++ backend/app/models/stage.py
@@ -101,7 +101,7 @@ class ProductBOMItem(Base):
     id = Column(Integer, primary_key=True, index=True)
     product_id = Column(Integer, ForeignKey("products.id"), nullable=False, index=True)
     material_id = Column(Integer, ForeignKey("materials.id"), nullable=False, index=True)
-    qty_per_unit = Column(Float, nullable=False)
+    qty_per_unit = Column(Numeric(10, 2), nullable=False)
     
     # Relationships
     product = relationship("Product", back_populates="bom_items")
